LayoutTests/sputnik/Conformance/08_Types/8.4_The_String_Type/S8.4_A13_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected string literal ''. Expected ';' after variable declaration.
S8.4_A13_T3

PASS Expected parsing failure

TEST COMPLETE
